queries rápidas!

+++ rebuild_database.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 import sqlite3
-import os
+import os, os.path
 import sys
 
 
@@ -15,42 +15,42 @@ if __name__ == "__main__":
         print('ANTES DE EXECUTAR: descompacte o arquivo ./dados/dados.zip no diretorio ./dados')
         print('Sem arquivos em dados/*.csv o banco não pode ser populado')
         sys.exit(1)
-    try:
-        os.remove('dados.db')
-    except:
-        print('Nao foi possivel deletar dados.db. Certifique-se que nao esta em uso ou que o arquivo exista')
+    if os.path.exists('dados.db'):
+        try:
+            os.remove('dados.db')
+        except:
+            print('Nao foi possivel deletar dados.db. Certifique-se que nao esta em uso ou que o arquivo exista')
         sys.exit(1)
     con = sqlite3.connect('dados.db')
     con.text_factory = str
     c = con.cursor()
-    c.execute('DROP TABLE IF EXISTS ACIDENTES')
-    c.execute('''CREATE TABLE ACIDENTES (
-                    VIA text,
-                    LOCAL text,     -- Logradouro, cruzamento, etc
-                    TIPO_ACID text, -- Abalroamento, Atropelamento, etc
-                    DIA_SEM text,    -- Sexta-feira, etc
-                    DIA int,
-                    MES int,
-                    ANO int,
-                    LATLONG text,    -- (precessar) latitude;longitude
+    c.execute('''CREATE TABLE acidentes (
+                    via text,
+                    local text,     -- Logradouro, cruzamento, etc
+                    tipo_acid text, -- Abalroamento, Atropelamento, etc
+                    dia_sem text,    -- Sexta-feira, etc
+                    dia int,
+                    mes int,
+                    ano int,
+                    latlong text,    -- (precessar) latitude;longitude
                 -- contagens
-                    FERIDOS int,    -- FERIDOS + FERIDOS_GR
-                    MORTES int,     -- MORTES + MORTES_POST
+                    feridos int,    -- FERIDOS + FERIDOS_GR
+                    mortes int,     -- MORTES + MORTES_POST
                 -- booleanos
-                    FATAL int,      -- FATAIS
-                    AUTO int,       -- automovel
-                    TAXI int,
-                    LOTACAO int,
-                    ONIBUS int,     -- ONIBUS_URB
-                    CAMINHAO int,
-                    MOTO int,
-                    CARROCA int,
-                    BICICLETA int,
-                    NOITE int       -- 0.dia 1.noite
+                    fatal int,      -- FATAIS
+                    auto int,       -- automovel
+                    taxi int,
+                    lotacao int,
+                    onibus int,     -- ONIBUS_URB
+                    caminhao int,
+                    moto int,
+                    carroca int,
+                    bicicleta int,
+                    noite int       -- 0.dia 1.noite
                 );
     ''')
-    c.execute('CREATE INDEX VIAINDEX ON ACIDENTES (VIA);');
-    c.execute('CREATE INDEX ANOINDEX ON ACIDENTES (ANO);');
+    c.execute('CREATE INDEX viaindex ON acidentes(via);');
+    c.execute('CREATE INDEX anoindex ON acidentes(ano);');
     
     for filename in os.listdir('%s/dados' % os.getcwd()):
         if filename.endswith('.csv'):
@@ -64,12 +64,10 @@ if __name__ == "__main__":
                     if len(row) == len(headers):
                         row = [x.replace(",",".") for x in row]
                         data.append([
-                            row[headers.index('LOG1')].capitalize() + (
-                                ' & %s' % row[headers.index('LOG2')].capitalize() if len(row[headers.index('LOG2')])>1 else ''
-                            ),
-                            row[headers.index('LOCAL')].capitalize(),
-                            row[headers.index('TIPO_ACID')].capitalize(),
-                            row[headers.index('DIA_SEM')].capitalize(),
+                            row[headers.index('LOG1')].title(),
+                            row[headers.index('LOCAL')].title(),
+                            row[headers.index('TIPO_ACID')].title(),
+                            row[headers.index('DIA_SEM')].title(),
                             int(row[headers.index('DIA')]),
                             int(row[headers.index('MES')]),
                             int(row[headers.index('ANO')]),
@@ -93,7 +91,7 @@ if __name__ == "__main__":
                             1 if row[headers.index('NOITE_DIA')]=='NOITE' else 0,
                         ])
 
-                sql = '''INSERT INTO ACIDENTES VALUES (%s)''' % (",".join(['?']*len(data[0])))
+                sql = '''INSERT INTO acidentes VALUES (%s)''' % (",".join(['?']*len(data[0])))
                 c.executemany(sql, data)
                 con.commit()
